CRUD`s complete

+++ src/app/admin/user-group/user-group.component.ts
@@ -1,17 +1,41 @@
-import { Component, OnInit, Input } from '@angular/core';
+import {Component, OnInit, Input, Inject} from '@angular/core';
 import {ApiService} from '../../services/api.service';
 import {ISubscription} from 'rxjs/Subscription';
+import {NgForm} from '@angular/forms';
+import {DOCUMENT} from '@angular/common';
+import {animate, state, style, transition, trigger} from '@angular/animations';
 
 @Component({
     selector: 'app-user-group',
-    templateUrl: './user-group.component.html'
+    templateUrl: './user-group.component.html',
+    styleUrls: ['../admin.component.css'],
+    animations: [
+        trigger('list', [
+            state('in', style({
+                opacity: 1,
+                transform: 'translateX(0)'
+            })),
+            transition('void => *', [
+                style({
+                    opacity: 0,
+                    transform: 'translateX(-100px)'
+                }),
+                animate(300)
+            ]),
+            transition('* => void', [
+                animate(300, style({
+                    opacity: 0,
+                    transform: 'translateX(100px)'
+                }))
+            ]),
+        ])
+    ]
 })
 export class UserGroupComponent implements OnInit {
-    @Input() group;
-
+    edit;
     groups;
     private groupsSubscription: ISubscription;
-    constructor(private apiService: ApiService) {
+    constructor(private apiService: ApiService, @Inject(DOCUMENT) private doc: Document) {
         this.groupsSubscription = this.apiService.getUserGroups().subscribe(
             (groups) => {
                 console.log(groups);
@@ -22,22 +46,25 @@ export class UserGroupComponent implements OnInit {
 
     ngOnInit() {
     }
-
-    deleteUserGroup(id: number) {
-        this.apiService.deleteUserGroup(id)
-            .subscribe(
-                (groups) => {
-                    console.log(groups);
-                }
-            );
-        this.groups = this.groups.filter( group => group.id !== id);
+    remove(i: number) {
+        const id = this.groups[i].id;
+        if(confirm('Are you sure to delete: ' + this.groups[i].title)) {
+            this.apiService.deleteUserGroup(id)
+                .subscribe(
+                    (group) => {
+                        console.log(group);
+                    }
+                );
+            this.groups = this.groups.filter( status => status.id !== id);
+        }
     }
 
-    updateUserGroup(id) {
+    showUpdate(i) {
+        this.groups[i].edit = true;
     }
 
-    createUserGroup(title: string) {
-        this.apiService.createUserGroup(title).subscribe(
+    create(form: NgForm) {
+        this.apiService.createUserGroup(form.value.title).subscribe(
             (group) => {
                 console.log(group);
                 this.groups.push({id: group.id, title: group.title});
@@ -45,13 +72,19 @@ export class UserGroupComponent implements OnInit {
         );
     }
 
+    cancel(i: number) {
+        const id = this.groups[i].id;
+        (<HTMLInputElement>this.doc.getElementById('input-title-' + id)).value = this.groups[i].title;
+        this.groups[i].edit = false;
+    }
 
-    // Save(title: string) {
-    //     this.apiService.updateStatus(title).subscribe(
-    //         (role) => {
-    //             console.log(role);
-    //             this.groups.push({id: role.id, title: role.title});
-    //         }
-    //     );
-    // }
+    update(form: NgForm, i: number) {
+        const id = this.groups[i].id;
+        this.apiService.updateUserGroup(id, form.value.title).subscribe(
+            (group) => {
+                this.groups[i].title = group.title;
+                this.groups[i].edit = false;
+            }
+        );
+    }
 }
